+++ index.php
@@ -1,7 +1,7 @@
 <?php
 require_once __DIR__ . '\vendor\autoload.php';
 
-use src\AvailableActions;
+use src\logic\AvailableActions;
 
 $task = new AvailableActions(1,
     'String',
@@ -10,6 +10,6 @@ $task = new AvailableActions(1,
     '2019-11-03',
     2,);
 echo '<pre>';
-print_r($task->getAvailableActions(AvailableActions::ROLE_EXECUTOR));
+print_r($task->getAvailableActions(AvailableActions::ROLE_CUSTOMER));
