+++ src/logic/AvailableActions.php
@@ -1,7 +1,7 @@
 <?php
 
 
-namespace src;
+namespace src\logic;
 
 
 class AvailableActions
